% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfdown-package.R
\docType{package}
\name{tfdown-package}
\alias{tfdown}
\alias{tfdown-package}
\title{tfdown: Yet Another Tensorflow For R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/tfdown}
  \item Report bugs at \url{https://github.com/mrchypark/tfdown/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
